wip(checkpoint): effective-state-test-retirement as reviewed on 2026-07-16 base be3e93ce
Checkpoint of the test-retirement slice exactly as it passed its bounded
review at be3e93ce (tasks/reviews/20260716-0222-effective-state-test-retirement.review.md,
recommendation pass). Not merged: the review's benchmark-evidence freshness
blocker still stands, and post-LSC main has since rewritten the retired test
surfaces (effective-state, adapter-parity, cli-state-golden, project-effective-state).
Needs rebase reconciliation onto post-LSC main as its own bounded slice.
